requirements.txt
speedtest-cli
bitmath==1.0.2-3